+++ tests/lib/rules/no-multi-spaces.js
@@ -73,6 +73,38 @@ ruleTester.run('no-multi-spaces', rule, {
           ignoreProperties: true
         }
       ]
+    },
+    {
+      code: `
+      <template>
+        <div
+          :class="{
+           'foo': foo,   // comment
+          }"
+        ></div>
+      </template>
+      `,
+      options: [
+        {
+          ignoreEOLComments: true
+        }
+      ]
+    },
+    {
+      code: `
+      <template>
+        <div
+          :class="{
+            'foo': foo   /* multiline comment */
+          }"
+        ></div>
+      </template>
+      `,
+      options: [
+        {
+          ignoreEOLComments: true
+        }
+      ]
     }
   ],
   invalid: [
@@ -82,7 +114,6 @@ ruleTester.run('no-multi-spaces', rule, {
       errors: [
         {
           message: "Multiple spaces found before '/>'.",
-          type: 'HTMLSelfClosingTagClose',
           line: 1,
           column: 15,
           endLine: 1,
@@ -96,7 +127,6 @@ ruleTester.run('no-multi-spaces', rule, {
       errors: [
         {
           message: "Multiple spaces found before 'class'.",
-          type: 'HTMLIdentifier',
           line: 1,
           column: 15,
           endLine: 1,
@@ -104,7 +134,6 @@ ruleTester.run('no-multi-spaces', rule, {
         },
         {
           message: "Multiple spaces found before '/>'.",
-          type: 'HTMLSelfClosingTagClose',
           line: 1,
           column: 29,
           endLine: 1,
@@ -118,7 +147,6 @@ ruleTester.run('no-multi-spaces', rule, {
       errors: [
         {
           message: "Multiple spaces found before 'class'.",
-          type: 'HTMLIdentifier',
           line: 1,
           column: 15,
           endLine: 1,
@@ -126,7 +154,6 @@ ruleTester.run('no-multi-spaces', rule, {
         },
         {
           message: "Multiple spaces found before '/>'.",
-          type: 'HTMLSelfClosingTagClose',
           line: 1,
           column: 28,
           endLine: 1,
@@ -140,7 +167,6 @@ ruleTester.run('no-multi-spaces', rule, {
       errors: [
         {
           message: "Multiple spaces found before ':'.",
-          type: 'Punctuator',
           line: 1,
           column: 15,
           endLine: 1,
@@ -148,7 +174,6 @@ ruleTester.run('no-multi-spaces', rule, {
         },
         {
           message: "Multiple spaces found before '/>'.",
-          type: 'HTMLSelfClosingTagClose',
           line: 1,
           column: 30,
           endLine: 1,
@@ -162,7 +187,6 @@ ruleTester.run('no-multi-spaces', rule, {
       errors: [
         {
           message: "Multiple spaces found before '/>'.",
-          type: 'HTMLSelfClosingTagClose',
           line: 1,
           column: 35,
           endLine: 1,
@@ -176,7 +200,6 @@ ruleTester.run('no-multi-spaces', rule, {
       errors: [
         {
           message: "Multiple spaces found before '/>'.",
-          type: 'HTMLSelfClosingTagClose',
           line: 1,
           column: 34,
           endLine: 1,
@@ -190,7 +213,6 @@ ruleTester.run('no-multi-spaces', rule, {
       errors: [
         {
           message: "Multiple spaces found before '/>'.",
-          type: 'HTMLSelfClosingTagClose',
           line: 1,
           column: 36,
           endLine: 1,
@@ -204,7 +226,6 @@ ruleTester.run('no-multi-spaces', rule, {
       errors: [
         {
           message: "Multiple spaces found before '/>'.",
-          type: 'HTMLSelfClosingTagClose',
           line: 2,
           column: 21,
           endLine: 2,
@@ -218,7 +239,6 @@ ruleTester.run('no-multi-spaces', rule, {
       errors: [
         {
           message: "Multiple spaces found before 'test'.",
-          type: 'Identifier',
           line: 1,
           column: 18,
           endLine: 1,
@@ -226,7 +246,6 @@ ruleTester.run('no-multi-spaces', rule, {
         },
         {
           message: "Multiple spaces found before '}}'.",
-          type: 'VExpressionEnd',
           line: 1,
           column: 24,
           endLine: 1,
@@ -240,7 +259,6 @@ ruleTester.run('no-multi-spaces', rule, {
       errors: [
         {
           message: "Multiple spaces found before '>'.",
-          type: 'HTMLTagClose',
           line: 1,
           column: 15,
           endLine: 1,
@@ -254,7 +272,6 @@ ruleTester.run('no-multi-spaces', rule, {
       errors: [
         {
           message: "Multiple spaces found before 'i'.",
-          type: 'Identifier',
           line: 1,
           column: 23,
           endLine: 1,
@@ -262,7 +279,6 @@ ruleTester.run('no-multi-spaces', rule, {
         },
         {
           message: "Multiple spaces found before 'in'.",
-          type: 'Keyword',
           line: 1,
           column: 30,
           endLine: 1,
@@ -270,7 +286,6 @@ ruleTester.run('no-multi-spaces', rule, {
         },
         {
           message: "Multiple spaces found before 'b'.",
-          type: 'Identifier',
           line: 1,
           column: 36,
           endLine: 1,
@@ -278,7 +293,6 @@ ruleTester.run('no-multi-spaces', rule, {
         },
         {
           message: "Multiple spaces found before '\"'.",
-          type: 'Punctuator',
           line: 1,
           column: 41,
           endLine: 1,
@@ -310,7 +324,6 @@ ruleTester.run('no-multi-spaces', rule, {
       errors: [
         {
           message: "Multiple spaces found before ':'.",
-          type: 'Punctuator',
           line: 5,
           column: 26,
           endLine: 5,
@@ -342,11 +355,136 @@ ruleTester.run('no-multi-spaces', rule, {
       errors: [
         {
           message: "Multiple spaces found before 'isExpanded'.",
-          type: 'Identifier',
           line: 5,
           column: 27,
           endLine: 5,
           endColumn: 30
+        }
+      ]
+    },
+    {
+      code: `
+      <template>
+        <div
+          :class="{
+            'foo': foo  // comment
+          }"
+        ></div>
+      </template>
+      `,
+      output: `
+      <template>
+        <div
+          :class="{
+            'foo': foo // comment
+          }"
+        ></div>
+      </template>
+      `,
+      errors: [
+        {
+          message: "Multiple spaces found before '// comment'.",
+          line: 5,
+          column: 23,
+          endLine: 5,
+          endColumn: 25
+        }
+      ]
+    },
+    {
+      code: `
+      <template>
+        <div
+          :class="{
+            'foo': foo  /* multiline comment */
+          }"
+        ></div>
+      </template>
+      `,
+      output: `
+      <template>
+        <div
+          :class="{
+            'foo': foo /* multiline comment */
+          }"
+        ></div>
+      </template>
+      `,
+      errors: [
+        {
+          message: "Multiple spaces found before '/* multiline comment */'.",
+          line: 5,
+          column: 23,
+          endLine: 5,
+          endColumn: 25
+        }
+      ]
+    },
+    {
+      code: `
+      <template>
+        <div
+          :class="{
+            'foo': foo  // comment
+          }"
+        ></div>
+      </template>
+      `,
+      output: `
+      <template>
+        <div
+          :class="{
+            'foo': foo // comment
+          }"
+        ></div>
+      </template>
+      `,
+      options: [
+        {
+          ignoreEOLComments: false
+        }
+      ],
+      errors: [
+        {
+          message: "Multiple spaces found before '// comment'.",
+          line: 5,
+          column: 23,
+          endLine: 5,
+          endColumn: 25
+        }
+      ]
+    },
+    {
+      code: `
+      <template>
+        <div
+          :class="{
+            'foo': foo  /* multiline comment */
+          }"
+        ></div>
+      </template>
+      `,
+      output: `
+      <template>
+        <div
+          :class="{
+            'foo': foo /* multiline comment */
+          }"
+        ></div>
+      </template>
+      `,
+      options: [
+        {
+          ignoreEOLComments: false
+        }
+      ],
+      errors: [
+        {
+          message: "Multiple spaces found before '/* multiline comment */'.",
+          line: 5,
+          column: 23,
+          endLine: 5,
+          endColumn: 25
         }
       ]
     }