+++ tests/lib/rules/no-reserved-component-names.js
@@ -608,8 +608,10 @@ ruleTester.run('no-reserved-component-names', rule, {
             ? 'reservedInHtml'
             : 'reserved',
           data: { name },
-          type: 'Literal',
-          line: 3
+          line: 3,
+          column: 19,
+          endLine: 3,
+          endColumn: 19 + name.length + 2
         }
       ]
     })),
@@ -623,8 +625,10 @@ ruleTester.run('no-reserved-component-names', rule, {
             ? 'reservedInHtml'
             : 'reserved',
           data: { name },
-          type: 'Literal',
-          line: 1
+          line: 1,
+          column: 15,
+          endLine: 1,
+          endColumn: 15 + name.length + 2
         }
       ]
     })),
@@ -638,8 +642,10 @@ ruleTester.run('no-reserved-component-names', rule, {
             ? 'reservedInHtml'
             : 'reserved',
           data: { name },
-          type: 'Literal',
-          line: 1
+          line: 1,
+          column: 15,
+          endLine: 1,
+          endColumn: 15 + name.length + 2
         }
       ]
     })),
@@ -653,8 +659,10 @@ ruleTester.run('no-reserved-component-names', rule, {
             ? 'reservedInHtml'
             : 'reserved',
           data: { name },
-          type: 'TemplateLiteral',
-          line: 1
+          line: 1,
+          column: 15,
+          endLine: 1,
+          endColumn: 15 + name.length + 2
         }
       ]
     })),
@@ -668,8 +676,10 @@ ruleTester.run('no-reserved-component-names', rule, {
             ? 'reservedInHtml'
             : 'reserved',
           data: { name },
-          type: 'TemplateLiteral',
-          line: 1
+          line: 1,
+          column: 15,
+          endLine: 1,
+          endColumn: 15 + name.length + 2
         }
       ]
     })),
@@ -687,8 +697,10 @@ ruleTester.run('no-reserved-component-names', rule, {
             ? 'reservedInHtml'
             : 'reserved',
           data: { name },
-          type: 'Property',
-          line: 3
+          line: 3,
+          column: 13,
+          endLine: 3,
+          endColumn: 13 + name.length + 6
         }
       ]
     })),
@@ -705,7 +717,10 @@ ruleTester.run('no-reserved-component-names', rule, {
             ? 'reservedInHtml'
             : 'reserved',
           data: { name },
-          line: 1
+          line: 1,
+          column: 37,
+          endLine: 1,
+          endColumn: 37 + name.length + 2
         }
       ]
     })),
@@ -723,7 +738,10 @@ ruleTester.run('no-reserved-component-names', rule, {
             ? 'reservedInHtml'
             : 'reserved',
           data: { name },
-          line: 1
+          line: 1,
+          column: 37,
+          endLine: 1,
+          endColumn: 37 + name.length + 2
         }
       ]
     })),
@@ -740,8 +758,10 @@ ruleTester.run('no-reserved-component-names', rule, {
         {
           messageId: 'reservedInVue',
           data: { name },
-          type: 'Literal',
-          line: 3
+          line: 3,
+          column: 19,
+          endLine: 3,
+          endColumn: 19 + name.length + 2
         }
       ]
     })),
@@ -758,8 +778,10 @@ ruleTester.run('no-reserved-component-names', rule, {
         {
           messageId: 'reservedInVue',
           data: { name },
-          type: 'Literal',
-          line: 3
+          line: 3,
+          column: 19,
+          endLine: 3,
+          endColumn: 19 + name.length + 2
         }
       ]
     })),
@@ -776,8 +798,10 @@ ruleTester.run('no-reserved-component-names', rule, {
         {
           messageId: 'reservedInVue3',
           data: { name },
-          type: 'Literal',
-          line: 3
+          line: 3,
+          column: 19,
+          endLine: 3,
+          endColumn: 19 + name.length + 2
         }
       ]
     }))