+++ tests/lib/rules/prop-name-casing.js
@@ -4,7 +4,6 @@
  */
 'use strict'
 
-const semver = require('semver')
 const rule = require('../../../lib/rules/prop-name-casing')
 const RuleTester = require('../../eslint-compat').RuleTester
 const {
@@ -406,8 +405,10 @@ ruleTester.run('prop-name-casing', rule, {
       errors: [
         {
           message: 'Prop "greeting_text" is not in camelCase.',
-          type: 'Property',
-          line: 4
+          line: 4,
+          column: 13,
+          endLine: 4,
+          endColumn: 34
         }
       ]
     },
@@ -425,8 +426,10 @@ ruleTester.run('prop-name-casing', rule, {
       errors: [
         {
           message: 'Prop "greeting_text" is not in camelCase.',
-          type: 'Property',
-          line: 4
+          line: 4,
+          column: 13,
+          endLine: 4,
+          endColumn: 34
         }
       ]
     },
@@ -442,8 +445,10 @@ ruleTester.run('prop-name-casing', rule, {
       errors: [
         {
           message: 'Prop "greeting_text" is not in camelCase.',
-          type: 'Literal',
-          line: 3
+          line: 3,
+          column: 19,
+          endLine: 3,
+          endColumn: 34
         }
       ]
     },
@@ -461,8 +466,10 @@ ruleTester.run('prop-name-casing', rule, {
       errors: [
         {
           message: 'Prop "greetingText" is not in snake_case.',
-          type: 'Property',
-          line: 4
+          line: 4,
+          column: 13,
+          endLine: 4,
+          endColumn: 33
         }
       ]
     },
@@ -480,8 +487,10 @@ ruleTester.run('prop-name-casing', rule, {
       errors: [
         {
           message: 'Prop "greeting-text" is not in camelCase.',
-          type: 'Property',
-          line: 4
+          line: 4,
+          column: 13,
+          endLine: 4,
+          endColumn: 36
         }
       ]
     },
@@ -499,8 +508,10 @@ ruleTester.run('prop-name-casing', rule, {
       errors: [
         {
           message: 'Prop "greeting-text" is not in snake_case.',
-          type: 'Property',
-          line: 4
+          line: 4,
+          column: 13,
+          endLine: 4,
+          endColumn: 36
         }
       ]
     },
@@ -517,8 +528,10 @@ ruleTester.run('prop-name-casing', rule, {
       errors: [
         {
           message: 'Prop "greeting_text" is not in camelCase.',
-          type: 'Property',
-          line: 4
+          line: 4,
+          column: 13,
+          endLine: 4,
+          endColumn: 36
         }
       ]
     },
@@ -536,8 +549,10 @@ ruleTester.run('prop-name-casing', rule, {
       errors: [
         {
           message: 'Prop "greeting-text" is not in camelCase.',
-          type: 'Property',
-          line: 4
+          line: 4,
+          column: 13,
+          endLine: 4,
+          endColumn: 38
         }
       ]
     },
@@ -555,8 +570,10 @@ ruleTester.run('prop-name-casing', rule, {
       errors: [
         {
           message: 'Prop "greeting_text" is not in camelCase.',
-          type: 'Property',
-          line: 4
+          line: 4,
+          column: 13,
+          endLine: 4,
+          endColumn: 26
         }
       ]
     },
@@ -573,8 +590,10 @@ ruleTester.run('prop-name-casing', rule, {
       errors: [
         {
           message: 'Prop "abc-123-def" is not in camelCase.',
-          type: 'Property',
-          line: 4
+          line: 4,
+          column: 13,
+          endLine: 4,
+          endColumn: 34
         }
       ]
     },
@@ -592,8 +611,10 @@ ruleTester.run('prop-name-casing', rule, {
       errors: [
         {
           message: 'Prop "greeting-text" is not in camelCase.',
-          type: 'Property',
-          line: 4
+          line: 4,
+          column: 13,
+          endLine: 4,
+          endColumn: 40
         }
       ]
     },
@@ -664,7 +685,10 @@ ruleTester.run('prop-name-casing', rule, {
       errors: [
         {
           message: 'Prop "greeting_text" is not in camelCase.',
-          line: 4
+          line: 4,
+          column: 9,
+          endLine: 4,
+          endColumn: 30
         }
       ]
     },
@@ -682,41 +706,40 @@ ruleTester.run('prop-name-casing', rule, {
       errors: [
         {
           message: 'Prop "greeting_text" is not in camelCase.',
-          line: 3
+          line: 3,
+          column: 20,
+          endLine: 3,
+          endColumn: 35
         }
       ]
     },
-    ...(semver.lt(
-      require('@typescript-eslint/parser/package.json').version,
-      '4.0.0'
-    )
-      ? []
-      : [
-          {
-            filename: 'test.vue',
-            code: `
-            <script setup lang="ts">
-            interface Props {
-              greeting_text: number
-            }
-            defineProps<Props>()
-            </script>
-            `,
-            languageOptions: {
-              parser: require('vue-eslint-parser'),
-              ...languageOptions,
-              parserOptions: {
-                parser: require.resolve('@typescript-eslint/parser')
-              }
-            },
-            errors: [
-              {
-                message: 'Prop "greeting_text" is not in camelCase.',
-                line: 4
-              }
-            ]
-          }
-        ]),
+    {
+      filename: 'test.vue',
+      code: `
+      <script setup lang="ts">
+      interface Props {
+        greeting_text: number
+      }
+      defineProps<Props>()
+      </script>
+      `,
+      languageOptions: {
+        parser: require('vue-eslint-parser'),
+        ...languageOptions,
+        parserOptions: {
+          parser: require.resolve('@typescript-eslint/parser')
+        }
+      },
+      errors: [
+        {
+          message: 'Prop "greeting_text" is not in camelCase.',
+          line: 4,
+          column: 9,
+          endLine: 4,
+          endColumn: 30
+        }
+      ]
+    },
     {
       filename: 'test.vue',
       code: `
@@ -733,13 +756,17 @@ ruleTester.run('prop-name-casing', rule, {
       errors: [
         {
           message: 'Prop "notIgnored_prop" is not in camelCase.',
-          type: 'Property',
-          line: 4
+          line: 4,
+          column: 13,
+          endLine: 4,
+          endColumn: 36
         },
         {
           message: 'Prop "other-pattern" is not in camelCase.',
-          type: 'Property',
-          line: 5
+          line: 5,
+          column: 13,
+          endLine: 5,
+          endColumn: 36
         }
       ]
     },
@@ -755,13 +782,17 @@ ruleTester.run('prop-name-casing', rule, {
       errors: [
         {
           message: 'Prop "notIgnored_prop" is not in camelCase.',
-          type: 'Literal',
-          line: 3
+          line: 3,
+          column: 19,
+          endLine: 3,
+          endColumn: 36
         },
         {
           message: 'Prop "pattern_invalid" is not in camelCase.',
-          type: 'Literal',
-          line: 3
+          line: 3,
+          column: 38,
+          endLine: 3,
+          endColumn: 55
         }
       ]
     },
@@ -776,15 +807,24 @@ ruleTester.run('prop-name-casing', rule, {
       errors: [
         {
           message: 'Prop "snake_case" is not in camelCase.',
-          line: 5
+          line: 5,
+          column: 19,
+          endLine: 5,
+          endColumn: 24
         },
         {
           message: 'Prop "kebab-case" is not in camelCase.',
-          line: 5
+          line: 5,
+          column: 19,
+          endLine: 5,
+          endColumn: 24
         },
         {
           message: 'Prop "PascalCase" is not in camelCase.',
-          line: 5
+          line: 5,
+          column: 19,
+          endLine: 5,
+          endColumn: 24
         }
       ],
       ...getTypeScriptFixtureTestOptions()