+++ tests/lib/utils/vue-component.js
@@ -26,11 +26,13 @@ const languageOptions = {
   sourceType: 'module'
 }
 
-function makeError(line) {
+function makeError(line, column, endLine, endColumn) {
   return {
     message: 'Component detected.',
     line,
-    type: 'ObjectExpression'
+    column,
+    endLine,
+    endColumn
   }
 }
 
@@ -132,37 +134,37 @@ function invalidTests(ext) {
         // ${ext}
       `,
       languageOptions,
-      errors: [makeError(4)]
+      errors: [makeError(4, 19, 4, 21)]
     },
     {
       filename: `test.${ext}`,
       code: `Vue.component({})`,
       languageOptions,
-      errors: [makeError(1)]
+      errors: [makeError(1, 15, 1, 17)]
     },
     {
       filename: `test.${ext}`,
       code: `Vue.mixin({})`,
       languageOptions,
-      errors: [makeError(1)]
+      errors: [makeError(1, 11, 1, 13)]
     },
     {
       filename: `test.${ext}`,
       code: `Vue.extend({})`,
       languageOptions,
-      errors: [makeError(1)]
+      errors: [makeError(1, 12, 1, 14)]
     },
     {
       filename: `test.${ext}`,
       code: `app.component('name', {})`,
       languageOptions,
-      errors: [makeError(1)]
+      errors: [makeError(1, 23, 1, 25)]
     },
     {
       filename: `test.${ext}`,
       code: `app.mixin({})`,
       languageOptions,
-      errors: [makeError(1)]
+      errors: [makeError(1, 11, 1, 13)]
     },
     {
       filename: `test.${ext}`,
@@ -171,7 +173,7 @@ function invalidTests(ext) {
         ...languageOptions,
         parser: require('@typescript-eslint/parser')
       },
-      errors: [makeError(1)]
+      errors: [makeError(1, 52, 1, 54)]
     },
     {
       filename: `test.${ext}`,
@@ -180,7 +182,7 @@ function invalidTests(ext) {
         ...languageOptions,
         parser: require('@typescript-eslint/parser')
       },
-      errors: [makeError(1)]
+      errors: [makeError(1, 27, 1, 29)]
     },
     {
       filename: `test.${ext}`,
@@ -189,13 +191,13 @@ function invalidTests(ext) {
         ...languageOptions,
         parser: require('@typescript-eslint/parser')
       },
-      errors: [makeError(1)]
+      errors: [makeError(1, 27, 1, 29)]
     },
     {
       filename: `test.${ext}`,
       code: `createApp({})`,
       languageOptions,
-      errors: [makeError(1)]
+      errors: [makeError(1, 11, 1, 13)]
     },
     {
       filename: `test.${ext}`,
@@ -205,7 +207,7 @@ function invalidTests(ext) {
         // ${ext}
       `,
       languageOptions,
-      errors: [makeError(3)]
+      errors: [makeError(3, 24, 3, 27)]
     },
     {
       filename: `test.${ext}`,
@@ -215,7 +217,7 @@ function invalidTests(ext) {
         // ${ext}
       `,
       languageOptions,
-      errors: [makeError(3)]
+      errors: [makeError(3, 24, 3, 27)]
     },
     {
       filename: `test.${ext}`,
@@ -228,7 +230,7 @@ function invalidTests(ext) {
         // ${ext}
       `,
       languageOptions,
-      errors: [makeError(6)]
+      errors: [makeError(6, 24, 6, 27)]
     },
     {
       filename: `test.${ext}`,
@@ -240,7 +242,7 @@ function invalidTests(ext) {
         // ${ext}
       `,
       languageOptions,
-      errors: [makeError(3), makeError(5)]
+      errors: [makeError(3, 24, 3, 27), makeError(5, 24, 5, 27)]
     },
     {
       filename: `test.${ext}`,
@@ -252,7 +254,7 @@ function invalidTests(ext) {
         // ${ext}
       `,
       languageOptions,
-      errors: [makeError(3), makeError(5)]
+      errors: [makeError(3, 28, 3, 31), makeError(5, 24, 5, 27)]
     },
     {
       filename: `test.${ext}`,
@@ -263,7 +265,10 @@ function invalidTests(ext) {
         // ${ext}
       `,
       languageOptions,
-      errors: [...(ext === 'js' ? [] : [makeError(2)]), makeError(4)]
+      errors: [
+        ...(ext === 'js' ? [] : [makeError(2, 24, 2, 27)]),
+        makeError(4, 26, 4, 29)
+      ]
     },
     {
       filename: `test.${ext}`,
@@ -274,7 +279,7 @@ function invalidTests(ext) {
         // ${ext}
       `,
       languageOptions,
-      errors: [makeError(4)]
+      errors: [makeError(4, 26, 4, 29)]
     },
     {
       filename: `test.${ext}`,
@@ -286,7 +291,7 @@ function invalidTests(ext) {
         // ${ext}
       `,
       languageOptions,
-      errors: [makeError(4)]
+      errors: [makeError(4, 13, 4, 16)]
     },
     {
       filename: `test.${ext}`,
@@ -301,7 +306,10 @@ function invalidTests(ext) {
         // ${ext}
       `,
       languageOptions,
-      errors: [...(ext === 'js' ? [] : [makeError(3)]), makeError(6)]
+      errors: [
+        ...(ext === 'js' ? [] : [makeError(3, 24, 7, 10)]),
+        makeError(6, 16, 6, 19)
+      ]
     },
     {
       filename: `test.${ext}`,
@@ -318,19 +326,22 @@ function invalidTests(ext) {
         // ${ext}
       `,
       languageOptions,
-      errors: [...(ext === 'js' ? [] : [makeError(2)]), makeError(8)]
+      errors: [
+        ...(ext === 'js' ? [] : [makeError(2, 24, 10, 10)]),
+        makeError(8, 20, 8, 22)
+      ]
     },
     {
       filename: `test.${ext}`,
       code: `export default defineComponent({})`,
       languageOptions,
-      errors: [makeError(1)]
+      errors: [makeError(1, 32, 1, 34)]
     },
     {
       filename: `test.${ext}`,
       code: `export default defineNuxtComponent({})`,
       languageOptions,
-      errors: [makeError(1)]
+      errors: [makeError(1, 36, 1, 38)]
     }
   ]
 }
@@ -353,13 +364,13 @@ ruleTester.run('vue-component', rule, {
       filename: 'test.vue',
       code: `export default { }`,
       languageOptions,
-      errors: [makeError(1)]
+      errors: [makeError(1, 16, 1, 19)]
     },
     {
       filename: 'test.jsx',
       code: `export default { }`,
       languageOptions,
-      errors: [makeError(1)]
+      errors: [makeError(1, 16, 1, 19)]
     },
     ...invalidTests('js'),
     ...invalidTests('jsx'),